(multithreaded-blocking-service)Implementing multi threaded blocking server
This is very similar to blocking server, but it executes the handle method in a separate thread pool. So this can handle many requests, where as the single threader server hungs up on single server. There will be a limit on number of threads that can be spawned per process, So thread pool limit cannot be huge